+++ apps/automation/src/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AutomationEventConsumerModule } from './consumer/automation-event-consumer.module';
+import { AmqpEventConsumerModule } from './amqp/amqp-event-consumer.module';
 
 @Module({
   imports: [
@@ -9,6 +10,7 @@ import { AutomationEventConsumerModule } from './consumer/automation-event-consu
       envFilePath: ['.env.local', '.env', '.env.example'],
     }),
     AutomationEventConsumerModule,
+    AmqpEventConsumerModule,
   ],
 })
 export class AppModule {}