+++ apps/search/src/consumer/search-event-consumer.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { SearchEventConsumerService } from './search-event-consumer.service';
+import { AmqpEventConsumerService } from '../amqp/amqp-event-consumer.service';
 
 @Module({
-  providers: [SearchEventConsumerService],
+  providers: [SearchEventConsumerService, AmqpEventConsumerService],
 })
 export class SearchEventConsumerModule {}